requirements.txt
# FastAPI & Server
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.4

# OCR & PDF Processing
pdfplumber==0.11.0
pytesseract==0.3.10
Pillow==10.2.0

# LLM 
groq==0.4.2

# Database & Storage
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
boto3==1.34.69

# Utilities
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.1.1
boto3==1.34.48
python-multipart==0.0.9
pdfplumber==0.11.0
pytesseract==0.3.10
groq==0.4.2
pillow
